scan-email-attachments/pyproject.toml
[project]
name = "email-attachments-demo"
version = "0.1.0"
description = "DataXray Email Security Integration Demo - SMTP proxy with attachment classification"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtpd>=1.4.3",
    "dxrpy>=0.2.12",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "requests>=2.28.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["src/smtp_proxy"]
